Move midi playback logic into dedicated fragment
